backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.1
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
PyJWT==2.10.1
pyTelegramBotAPI==4.14.0
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
psycopg2-binary==2.9.9
